+++ internal/files/dirty.go
@@ -1,44 +1,82 @@
 package files
 
 import (
-	"errors"
-	"os"
-	"path/filepath"
-
 	cmdtoolkit "github.com/majohn-r/cmd-toolkit"
 	"github.com/majohn-r/output"
-	"github.com/spf13/afero"
 )
 
 const (
 	dirtyFileName = "metadata.dirty"
 )
 
+var (
+	stateFileInitializationFailureLogged bool
+	// a variable so testing can substitute another implementation
+	initStateFile = cmdtoolkit.InitStateFile
+)
+
+// MarkDirty mark the system dirty
 func MarkDirty(o output.Bus) {
-	fs := cmdtoolkit.FileSystem()
-	f := dirtyPath()
-	if _, fileErr := fs.Stat(f); fileErr != nil && errors.Is(fileErr, os.ErrNotExist) {
-		// ignore error - if the file didn't exist a moment ago, there is no
-		// reason to assume that the file cannot be written to
-		_ = afero.WriteFile(fs, f, []byte("dirty"), cmdtoolkit.StdFilePermissions)
-		o.Log(output.Info, "metadata dirty file written", map[string]any{"fileName": f})
+	sf := safeStateFile(o)
+	defer sf.Close()
+	if err := sf.Create(dirtyFileName); err != nil {
+		o.Log(output.Warning, "error creating dirty flag", map[string]any{"error": err})
+	} else {
+		o.Log(output.Info, "metadata dirty file created", map[string]any{"fileName": dirtyFileName})
 	}
 }
 
+// ClearDirty clears the system dirty state
 func ClearDirty(o output.Bus) {
-	f := dirtyPath()
-	if !cmdtoolkit.PlainFileExists(f) {
-		return
+	sf := safeStateFile(o)
+	defer sf.Close()
+	if err := sf.Remove(dirtyFileName); err != nil {
+		o.Log(output.Warning, "error removing dirty flag", map[string]any{"error": err})
+	} else {
+		o.Log(output.Info, "metadata dirty file deleted", map[string]any{"fileName": dirtyFileName})
 	}
-	// best effort
-	_ = cmdtoolkit.FileSystem().Remove(f)
-	o.Log(output.Info, "metadata dirty file deleted", map[string]any{"fileName": f})
 }
 
-func Dirty() bool {
-	return cmdtoolkit.PlainFileExists(dirtyPath())
+// Dirty returns whether the system is dirty
+func Dirty(o output.Bus) bool {
+	sf := safeStateFile(o)
+	defer sf.Close()
+	return sf.Exists(dirtyFileName)
 }
 
-func dirtyPath() string {
-	return filepath.Join(cmdtoolkit.ApplicationPath(), dirtyFileName)
+func safeStateFile(o output.Bus) cmdtoolkit.StateFile {
+	if sf, err := initStateFile("mp3repair"); err != nil || sf == nil {
+		if !stateFileInitializationFailureLogged {
+			o.Log(output.Warning, "cannot initialize directory for dirty flag", map[string]any{"error": err})
+			stateFileInitializationFailureLogged = true
+		}
+		return emergencyStateFile{}
+	} else {
+		return sf
+	}
+}
+
+type emergencyStateFile struct{}
+
+func (e emergencyStateFile) Read(_ string) ([]byte, error) {
+	return []byte{}, nil
+}
+
+func (e emergencyStateFile) Write(_ string, _ []byte) error {
+	return nil
+}
+
+func (e emergencyStateFile) Exists(_ string) bool {
+	return false
+}
+
+func (e emergencyStateFile) Create(_ string) error {
+	return nil
+}
+
+func (e emergencyStateFile) Remove(_ string) error {
+	return nil
+}
+
+func (e emergencyStateFile) Close() {
 }